Added speciation_dashboard.html
This is our current dataset of "survivor" genomes, 13-dimensional data plotted onto a 3D interactive grid. Uses unsupervised clustering to categorize micro-transformer genomes into "species." Includes easy single click on any data point to copy the genome's hash to clipboard, or click the species on the sidebar to copy all hashes for all individuals in that species, separated by commas, to the clipboard. This will, after the full project is uploaded, work in tandem with our testing scripts.